new csdl

+++ src/GUI/TTCTsp.java
@@ -1,43 +1,46 @@
 package GUI;
-
+import dto.currentuser;
 import javax.swing.*;
 
 import bll.SanPhamBLL;
 
 import java.awt.*;
+import java.awt.event.FocusEvent;
+import java.awt.event.FocusListener;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import dal.dbconnection;
+import dto.SanPhamDTO;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 
 public class TTCTsp extends BaseFrame {
-	private SanPhamBLL bllSanPham = new SanPhamBLL();
-	private JTextArea txtTenSP, txtMaSP, cbLoaiSP, txtSoLuong, txtHSX, txtQcdg, txtSolo, txtSodk, txtHSD, txtGia, txtMaLSP, txtMaHSX;
+    private SanPhamBLL bllSanPham = new SanPhamBLL();
+    private JTextArea txtTenSP, txtMaSP, cbLoaiSP, txtSoLuong, txtHSX, txtQcdg,  txtSodk,  txtGia, txtMaLSP, txtMaHSX;
     private JLabel lblInfo;
     public TTCTsp() {
-    	super("Thông tin sản phẩm");
-    	initialize();
-	} 
+        super("Thông tin sản phẩm");
+        initialize();
+    }
     @Override
-	protected void initUniqueComponents() {
-		 for (JButton btn : menuButtons) {
-	            if (btn.getText().equals("Sản phẩm")) {
-	                btn.setBackground(Color.decode("#EF5D7A")); 
-	                btn.setFont(new Font("Arial", Font.BOLD, 14)); 
-	            }
-	        }
-	        
-	        // Các nút khác vẫn giữ màu mặc định
-	        for (JButton btn : menuButtons) {
-	            if (!btn.getText().equals("Sản phẩm")) {
-	                btn.setBackground(Color.decode("#641A1F")); // Màu mặc định cho các nút khác
-	                btn.setFont(new Font("Arial", Font.BOLD, 12));
-	            }
-	        }
+    protected void initUniqueComponents() {
+        for (JButton btn : menuButtons) {
+            if (btn.getText().equals("Sản phẩm")) {
+                btn.setBackground(Color.decode("#EF5D7A"));
+                btn.setFont(new Font("Arial", Font.BOLD, 14));
+            }
+        }
+
+        // Các nút khác vẫn giữ màu mặc định
+        for (JButton btn : menuButtons) {
+            if (!btn.getText().equals("Sản phẩm")) {
+                btn.setBackground(Color.decode("#641A1F")); // Màu mặc định cho các nút khác
+                btn.setFont(new Font("Arial", Font.BOLD, 12));
+            }
+        }
 
 
         //Tiêu đề "Sản phẩm >> Thông tin sản phẩm" có thể nhấn
@@ -60,7 +63,7 @@ public class TTCTsp extends BaseFrame {
             }
         });
 
-       
+
 
         // Form nhập sản phẩm
         JLabel lblTenSP = new JLabel("Tên sản phẩm:");
@@ -73,7 +76,7 @@ public class TTCTsp extends BaseFrame {
         txtTenSP.setEditable(false);
         txtTenSP.setFont(new Font("Arial", Font.BOLD, 20));
         txtTenSP.setForeground(Color.decode("#641A1F"));
-        txtTenSP.setOpaque(false);  
+        txtTenSP.setOpaque(false);
         add(txtTenSP);
 
         JLabel lblMaSP = new JLabel("Mã sản phẩm:");
@@ -86,9 +89,9 @@ public class TTCTsp extends BaseFrame {
         txtMaSP.setEditable(false);
         txtMaSP.setFont(new Font("Arial", Font.BOLD, 20));
         txtMaSP.setForeground(Color.decode("#641A1F"));
-        txtMaSP.setOpaque(false); 
+        txtMaSP.setOpaque(false);
         add(txtMaSP);
-        
+
         JLabel lblLoaiSP = new JLabel("Loại sản phẩm:");
         lblLoaiSP.setBounds(270, 200, 150, 25);
         add(lblLoaiSP);
@@ -99,7 +102,7 @@ public class TTCTsp extends BaseFrame {
         cbLoaiSP.setEditable(false);
         cbLoaiSP.setFont(new Font("Arial", Font.BOLD, 20));
         cbLoaiSP.setForeground(Color.decode("#641A1F"));
-        cbLoaiSP.setOpaque(false);  
+        cbLoaiSP.setOpaque(false);
         add(cbLoaiSP);
 
         JLabel lblSoLuong = new JLabel("Số lượng:");
@@ -112,7 +115,7 @@ public class TTCTsp extends BaseFrame {
         txtSoLuong.setEditable(false);
         txtSoLuong.setFont(new Font("Arial", Font.BOLD, 20));
         txtSoLuong.setForeground(Color.decode("#641A1F"));
-        txtSoLuong.setOpaque(false);  
+        txtSoLuong.setOpaque(false);
         add(txtSoLuong);
 
         JLabel lblHSX = new JLabel("Hãng sản xuất:");
@@ -125,7 +128,7 @@ public class TTCTsp extends BaseFrame {
         txtHSX.setEditable(false);
         txtHSX.setFont(new Font("Arial", Font.BOLD, 20));
         txtHSX.setForeground(Color.decode("#641A1F"));
-        txtHSX.setOpaque(false);  
+        txtHSX.setOpaque(false);
         add(txtHSX);
 
         JLabel lblqcdg = new JLabel("Quy cách đóng gói:");
@@ -138,21 +141,9 @@ public class TTCTsp extends BaseFrame {
         txtQcdg.setEditable(false);
         txtQcdg.setFont(new Font("Arial", Font.BOLD, 20));
         txtQcdg.setForeground(Color.decode("#641A1F"));
-        txtQcdg.setOpaque(false);  
+        txtQcdg.setOpaque(false);
         add(txtQcdg);
 
-        JLabel lblSolo = new JLabel("Số lô:");
-        lblSolo.setBounds(270, 360, 150, 25);
-        add(lblSolo);
-        txtSolo = new JTextArea();
-        txtSolo.setBounds(270, 390, 300, 40);
-        txtSolo.setLineWrap(true);
-        txtSolo.setWrapStyleWord(true);
-        txtSolo.setEditable(false);
-        txtSolo.setFont(new Font("Arial", Font.BOLD, 20));
-        txtSolo.setForeground(Color.decode("#641A1F"));
-        txtSolo.setOpaque(false);  
-        add(txtSolo);
 
 
         JLabel lblSodk = new JLabel("Số đăng ký:");
@@ -165,21 +156,9 @@ public class TTCTsp extends BaseFrame {
         txtSodk.setEditable(false);
         txtSodk.setFont(new Font("Arial", Font.BOLD, 20));
         txtSodk.setForeground(Color.decode("#641A1F"));
-        txtSodk.setOpaque(false);  
+        txtSodk.setOpaque(false);
         add(txtSodk);
 
-        JLabel lblHSD = new JLabel("Hạn sử dụng:");
-        lblHSD.setBounds(270, 440, 150, 25);
-        add(lblHSD);
-        txtHSD = new JTextArea();
-        txtHSD.setBounds(270, 470, 300, 34);
-        txtHSD.setLineWrap(true);
-        txtHSD.setWrapStyleWord(true);
-        txtHSD.setEditable(false);
-        txtHSD.setFont(new Font("Arial", Font.BOLD, 20));
-        txtHSD.setForeground(Color.decode("#641A1F"));
-        txtHSD.setOpaque(false);  
-        add(txtHSD);
 
         JLabel lblGia = new JLabel("Giá bán:");
         lblGia.setBounds(700, 440, 150, 25);
@@ -191,9 +170,9 @@ public class TTCTsp extends BaseFrame {
         txtGia.setEditable(false);
         txtGia.setFont(new Font("Arial", Font.BOLD, 20));
         txtGia.setForeground(Color.decode("#641A1F"));
-        txtGia.setOpaque(false);  
+        txtGia.setOpaque(false);
         add(txtGia);
-        
+
         JButton btnXoa = new JButton("Xóa");
         btnXoa.setBounds(700, 520, 100, 40);
         btnXoa.setBackground(Color.decode("#F0483E"));
@@ -203,25 +182,28 @@ public class TTCTsp extends BaseFrame {
         btnXoa.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                int result = JOptionPane.showConfirmDialog(null, "Bạn có chắc muốn xóa sản phẩm này?", "Xác nhận xóa", JOptionPane.YES_NO_OPTION);
-                if (result == JOptionPane.YES_OPTION) {
-                    try (Connection conn = dbconnection.getConnection()) {
-                        String sql = "DELETE FROM SANPHAM WHERE MASP = ?";
-                        PreparedStatement pstmt = conn.prepareStatement(sql);
-                        pstmt.setString(1, txtMaSP.getText());
-                        int rowsAffected = pstmt.executeUpdate();
-                        if (rowsAffected > 0) {
-                            JOptionPane.showMessageDialog(null, "Đã xóa sản phẩm thành công!", "Thông báo", JOptionPane.INFORMATION_MESSAGE);
-                            dispose();
-                            new SanPham(); // trở lại giao diện danh sách sản phẩm
-                        } else {
-                            JOptionPane.showMessageDialog(null, "Xóa thất bại. Không tìm thấy sản phẩm!", "Lỗi", JOptionPane.ERROR_MESSAGE);
+                if (currentuser.coQuyen("XOASANPHAM")) {
+                    int result = JOptionPane.showConfirmDialog(null, "Bạn có chắc muốn xóa sản phẩm này?", "Xác nhận xóa", JOptionPane.YES_NO_OPTION);
+                    if (result == JOptionPane.YES_OPTION) {
+                        try (Connection conn = dbconnection.getConnection()) {
+                            String sql = "DELETE FROM SANPHAM WHERE MASP = ?";
+                            PreparedStatement pstmt = conn.prepareStatement(sql);
+                            pstmt.setString(1, txtMaSP.getText());
+                            int rowsAffected = pstmt.executeUpdate();
+                            if (rowsAffected > 0) {
+                                JOptionPane.showMessageDialog(null, "Đã xóa sản phẩm thành công!", "Thông báo", JOptionPane.INFORMATION_MESSAGE);
+                                dispose();
+                                new SanPham(); // trở lại giao diện danh sách sản phẩm
+                            } else {
+                                JOptionPane.showMessageDialog(null, "Xóa thất bại. Không tìm thấy sản phẩm!", "Lỗi", JOptionPane.ERROR_MESSAGE);
+                            }
+                        } catch (Exception ex) {
+                            ex.printStackTrace();
+                            JOptionPane.showMessageDialog(null, "Lỗi khi xóa sản phẩm!", "Lỗi", JOptionPane.ERROR_MESSAGE);
                         }
-                    } catch (Exception ex) {
-                        ex.printStackTrace();
-                        JOptionPane.showMessageDialog(null, "Lỗi khi xóa sản phẩm!", "Lỗi", JOptionPane.ERROR_MESSAGE);
-                    }
-                }
+                    }else {
+                        JOptionPane.showMessageDialog(null, "Bạn không có quyền xóa!", "Cảnh báo", JOptionPane.WARNING_MESSAGE);
+                    }}
             }
         });
 
@@ -230,44 +212,42 @@ public class TTCTsp extends BaseFrame {
         btnCapnhat.setBackground(Color.decode("#F0483E"));
         btnCapnhat.setForeground(Color.WHITE);
         add(btnCapnhat);
-        
+
         btnCapnhat.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
                 String maSP = txtMaSP.getText();  // hoặc nơi bạn hiển thị mã sản phẩm
+                dispose();
                 Capnhatttsp capNhatForm = new Capnhatttsp();
                 capNhatForm.loadProductInfoForUpdate(maSP);  // gọi phương thức load thông tin
                 capNhatForm.setVisible(true);  // mở form cập nhật
-                dispose();
             }
         });
 
 
         setVisible(true);
     }
-    
 
-    public void setThongTin(String tenSP, String maSP, String loaiSP, int soLuong, String maHSX, String dongGoi, 
-            String soLo, String soDk, String HSD, Double gia) {
-txtTenSP.setText(tenSP);
-txtMaSP.setText(maSP);
-cbLoaiSP.setText(loaiSP);
-txtSoLuong.setText(String.valueOf(soLuong));
-txtHSX.setText(maHSX);  // Sử dụng tham số maHSX
-txtQcdg.setText(dongGoi);
-txtSolo.setText(soLo);
-txtSodk.setText(soDk);
-txtHSD.setText(HSD);
-txtGia.setText(String.valueOf(gia));
 
-revalidate();
-repaint();
-}
+    public void setThongTin(String tenSP, String maSP, String loaiSP, int soLuong, String maHSX, String dongGoi,
+                            String soDk,  Double gia) {
+        txtTenSP.setText(tenSP);
+        txtMaSP.setText(maSP);
+        cbLoaiSP.setText(loaiSP);
+        txtSoLuong.setText(String.valueOf(soLuong));
+        txtHSX.setText(maHSX);  // Sử dụng tham số maHSX
+        txtQcdg.setText(dongGoi);
+        txtSodk.setText(soDk);
+        txtGia.setText(String.valueOf(gia));
+
+        revalidate();
+        repaint();
+    }
 
 
 
-     public static void main(String[] args) {
-         new TTCTsp();
-     }
+    public static void main(String[] args) {
+        new TTCTsp();
+    }
 }
